fix client alloc over flow

+++ src/args.h
@@ -49,7 +49,7 @@ typedef struct {
   uint16_t port;
   uint16_t mtu;
   uint16_t concurrency;
-	uint16_t clients;
+	uint32_t clients;
 	//in host order, linux srv start addres, and tun ip for client,
 	uint32_t netip;
   const char *up_script;
